+++ src/array.mjs
@@ -65,3 +65,4 @@ export { default as rsort } from './array/rsort.mjs';
 export { default as usort } from './array/usort.mjs';
 export { default as uksort } from './array/uksort.mjs';
 export { default as uasort } from './array/uasort.mjs';
+export { default as natsort } from './array/natsort.mjs';